+++ Sources/SuperwallKit/Paywall/Presentation/Internal Presentation/Operators/PresentPaywall.swift
@@ -30,19 +30,14 @@ extension Superwall {
     request: PresentationRequest,
     paywallStatePublisher: PassthroughSubject<PaywallState, Never>
   ) async throws {
-    Task.detached { [weak self] in
-      guard let self = self else {
-        return
-      }
-      let trackedEvent = InternalSuperwallEvent.PresentationRequest(
-        eventData: request.presentationInfo.eventData,
-        type: request.flags.type,
-        status: .presentation,
-        statusReason: nil,
-        factory: self.dependencyContainer
-      )
-      await self.track(trackedEvent)
-    }
+    let trackedEvent = InternalSuperwallEvent.PresentationRequest(
+      eventData: request.presentationInfo.eventData,
+      type: request.flags.type,
+      status: .presentation,
+      statusReason: nil,
+      factory: self.dependencyContainer
+    )
+    await self.track(trackedEvent)
 
     try await withCheckedThrowingContinuation { continuation in
       paywallViewController.present(